requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cassandra-driver==3.25.0
cffi==1.15.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.91.0
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.1
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
